Warning/Comment Cleanup
- CustomerObject.cpp
-- Cast doubles as Uint32 to resolve compile warning.

- DaySales.cpp
-- int changed to Uint32 in DaySales::Update() to resolve compile warning.

- Scene.h
-- Comment updates in SetPrices class to be more accurate about the class it's in.

+++ TrySomethingNew/GameObjects/CustomerObject.cpp
@@ -89,7 +89,7 @@ void CustomerObject::Update(Uint32 ticks) {
 			}
 		}
 
-		SetDrawPos(this->mPosition.GetX(), this->mPosition.GetY());
+		SetDrawPos((Uint32) this->mPosition.GetX(), (Uint32) this->mPosition.GetY());
 	}
 }
